NavBar added to landing page
