+++ app/page.tsx
@@ -1,109 +1,89 @@
-"use client";
+import Image from "next/image";
+import pattern from "../public/bg-pattern.svg";
+import gradient from "../public/purple-gradient.svg";
+import composition from "../public/purple-composition.svg";
 
-import React, { useState, useCallback } from "react";
-import { ChatWindow } from "@/components/ChatWindow";
+// navbar
+const Navbar = () => {
+  return (
+    <nav className="relative z-30 flex items-center justify-between mt-[30px] px-[80px] bg-transparent text-white">
+        <div className="flex justify-center align-middle">
+            <img src="/kaisen_logo_chat_window.svg" alt="kaisen-logo" />
+        </div>
+        <div className="flex flex-1 justify-evenly text-xl text-[#9C9C9C] px-32">
+            <a href="#features" className="hover:text-gray-300">
+            Features
+            </a>
+            <a href="#developers" className="hover:text-gray-300">
+            Developers
+            </a>
+            <a href="#blog" className="hover:text-gray-300">
+            Blog
+            </a>
+            <a href="#about" className="hover:text-gray-300">
+            About Us
+            </a>
+        </div>
+        <button 
+        className="px-6 py-2 rounded-lg text-white text-base font-medium hover:opacity-90 transition"
+        style={{
+            background: "linear-gradient(90deg, #8F59E2, #7321EB, #7E45D6)",
+          }}
+        >
+            Sign Up
+        </button>
+    </nav>
+  );
+};
 
 export default function Home() {
-  const userName: string = "John";
-  const walletAddress: string = "0x4f1f6cbf61a6c80b12dc...";
-
-  const currentTime: number = new Date().getHours();
-  const greeting: string = currentTime >= 17 ? "Good Evening" : "Good Morning";
-
-  // State to control the input and submission in ChatWindow
-  const [questionToSubmit, setQuestionToSubmit] = useState<string | null>(null);
-  // State to trigger InfoCard refresh
-  const [refreshKey, setRefreshKey] = useState(0);
-
-  const handleQuestionClick = useCallback((question: string) => {
-    setQuestionToSubmit(question);
-  }, []);
-
-  const handleNewChatClick = useCallback(() => {
-    setRefreshKey((prev) => prev + 1); // Increment to force InfoCard re-render
-  }, []);
-
-  const InfoCard: React.ReactElement = (
-    <section className="p-2 md:p-8 w-full max-h-[85%] overflow-hidden bg-transparent">
-      <div className="text-center mb-20">
-        <div className="flex justify-center">
-          <h3 className="flex flex-row text-4xl font-normal text-white max-[930px]:text-3xl max-[600px]:text-2xl">
-            {greeting}. <p className="text-[#A76BFF] ml-2.5">{userName}</p>
-          </h3>
-        </div>
-        <h2 className="text-6xl font-medium text-white max-[930px]:text-4xl max-[600px]:text-3xl">How can I help you?</h2>
-        <p className="text-2xl my-4 font-medium text-[#747474] max-[930px]:text-base max-[600px]:text-xs">
-          It all starts with a question. Ask anything from prices to trading
-          strategies. Kaisen turns your words into DeFi action.
-        </p>
-      </div>
-
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-4 mb-6 text-base">
-        <button
-          onClick={() => handleQuestionClick("What's the current price of $BTC and $APT?")}
-          className="p-3 rounded-lg border border-[#2A2F3C] text-[#E8EAED] hover:bg-[#2A2F3C] transition-colors max-[600px]:text-xs"
-          style={{
-            background: "linear-gradient(135deg, #2B2B2B, #151515, #202020)",
-          }}
-        >
-          What's the current price of $BTC and $APT?
-        </button>
-        <button
-          onClick={() => handleQuestionClick("Show me the top 5 gainers on Aptos today")}
-          className="p-3 rounded-lg border border-[#2A2F3C] text-[#E8EAED] hover:bg-[#2A2F3C] transition-colors max-[600px]:text-xs"
-          style={{
-            background: "linear-gradient(135deg, #2B2B2B, #151515, #202020)",
-          }}
-        >
-          Show me the top 5 gainers on Aptos today
-        </button>
-        <button
-          onClick={() => handleQuestionClick("Any trending tokens I should watch?")}
-          className="p-3 rounded-lg border border-[#2A2F3C] text-[#E8EAED] hover:bg-[#2A2F3C] transition-colors max-[600px]:text-xs"
-          style={{
-            background: "linear-gradient(135deg, #2B2B2B, #151515, #202020)",
-          }}
-        >
-          Any trending tokens I should watch?
-        </button>
-      </div>
-    </section>
-  );
-
   return (
-    <section className="flex h-screen bg-[url('/kaisen-background.svg')] bg-cover bg-no-repeat bg-center">
-      <aside className="w-1/4 bg-[#121212] p-4 overflow-y-auto border-r border-[#2A2F3C] opacity-70">
-        <div className="flex justify-center mb-12">
-          <img src="/kaisen_logo_chat_window.svg" alt="kaisen-logo-ChatHistory" />
-        </div>
-        <button
-          onClick={handleNewChatClick}
-          className="w-full mb-4 p-2 text-white rounded-[9.54px] text-base transition-colors hover:bg-[#5A3FE6]"
-          style={{
-            background: "linear-gradient(135deg, #8F59E2, #7321EB, #7E45D6)",
-          }}
-        >
-          + New Chat
-        </button>
-        <div className="flex flex-col gap-2">
-          <div className="p-2 bg-[#27303F] rounded-lg border border-[#2A2F3C] text-[#A7A7B8]">
-            {/* chat history */}
+    <>
+      <main className="relative max-h-screen overflow-hidden text-white bg-transparent">
+        <Image
+          src={pattern}
+          alt="Grid Pattern"
+          fill
+          className="absolute object-cover z-0 opacity-50"
+        />
+
+        <Image
+          src={gradient}
+          alt="Purple Gradient"
+          fill
+          className="absolute object-cover z-10"
+        />
+
+        <Image
+          src={composition}
+          alt="3D Composition"
+          width={600}
+          height={600}
+          className="absolute right-0 bottom-0 top-32 z-20 pointer-events-none mr-12 mb-12"
+        />
+
+        <Navbar />
+
+        <div className="relative z-30 flex flex-col items-start justify-center h-screen max-w-5xl ml-16 px-12 max-[768px]:px-6 max-[1024px]:px-10">
+          <h1 className="text-[106.5px] max-[1024px]:text-7xl max-[768px]:text-4xl font-medium leading-tight">
+            <span className="block">Talk DeFi.</span>
+            <span className="block text-white">Trade Smarter.</span>
+          </h1>
+
+          <p className="mt-6 text-[32px] max-[1024px]:text-2xl max-[768px]:text-base text-[#D4D4D4]">
+            Cut the noise. Use AI to lend, borrow, and trade — just by chatting.
+            Built on Aptos. Backed by real-time data.
+          </p>
+
+          <div className="flex justify-center items-center">
+            <button className="mt-10 px-12 py-3 text-2xl font-medium rounded-lg bg-gradient-to-r from-[#7B61FF] to-[#BA4EFF] hover:opacity-90 transition">
+              Connect your wallet
+            </button>
+          </div>
+          <div className="flex h-24">
           </div>
         </div>
-      </aside>
-
-      <main className="w-3/4 p-6">
-        <ChatWindow
-          endpoint="api/hello"
-          emoji="🤖"
-          titleText="KAISEN ASSISTANT"
-          placeholder="Ask about crypto trends and Twitter analysis..."
-          emptyStateComponent={InfoCard} // pass the static InfoCard
-          suggestedQuestion={questionToSubmit}
-          onQuestionSubmitted={() => setQuestionToSubmit(null)}
-          key={refreshKey} // force re-render of ChatWindow when refreshKey changes
-        />
       </main>
-    </section>
+    </>
   );
 }
